pyproject.toml
[build-system]
requires      = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name            = "chainvalidator"
version         = "0.1.0"
description     = "DNSSEC chain-of-trust validator – CLI utility and Python library"
readme          = "README.md"
requires-python = ">=3.11"
license         = "GPLv3"
authors         = [{ name = "t0kubetsu" }]
keywords        = ["dnssec", "dns", "security", "chain-of-trust", "validator", "cli"]
classifiers     = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: Name Service (DNS)",
    "Topic :: Security",
    "Topic :: System :: Networking",
    "Typing :: Typed",
]
dependencies    = [
    "dnspython[dnssec]>=2.6",
    "requests>=2.31",
    "rich>=13.7",
    "typer>=0.12",
]

[project.urls]
Homepage       = "https://github.com/yourname/chainvalidator"
Repository     = "https://github.com/yourname/chainvalidator"
Issues         = "https://github.com/yourname/chainvalidator/issues"
# Changelog      = "https://github.com/yourname/chainvalidator/blob/main/CHANGELOG.md"

[project.scripts]
chainvalidator = "chainvalidator.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
]

[tool.setuptools.packages.find]
where   = ["."]
include = ["chainvalidator*"]

[tool.setuptools.package-data]
chainvalidator = ["py.typed"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths  = ["tests"]